[clang-repl] Rework layering of incremental executors.
The original Interpreter implementation had a hard dependency on ORC and grew
organically with the addition of out-of-process JIT support. This tightly
coupled the Interpreter to a specific execution engine and leaked ORC-specific
assumptions (runtime layout, symbol lookup, exception model) into higher
layers.

The WebAssembly integration demonstrated that incremental execution can be
implemented without ORC, exposing the need for a cleaner abstraction boundary.

This change introduces an IncrementalExecutor interface and moves ORC-based
execution behind a concrete implementation. The Interpreter now depends only
on the abstract executor, improving layering and encapsulation.

In addition, the Interpreter can be configured with user-provided incremental
executor implementations, enabling ORC-independent execution, easier testing,
and future extensions without modifying the core Interpreter.
